Started refactoring client by moving kernel specific code from RPC lib into the kernel
